app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
neo4j
python-dotenv
